requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==8.0.0
django-countries==7.5.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-timezone-field==5.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.2
h11==0.14.0
idna==3.4
kombu==5.3.1
MarkupPy==1.14
mysql-connector-python==8.1.0
node==1.2.1
npm==0.1.1
oauthlib==3.2.2
odfpy==1.4.1
odict==1.9.0
openpyxl==3.1.2
optional-django==0.1.0
outcome==1.2.0
pbr==5.11.1
phonenumbers==8.13.15
pika==1.3.2
Pillow==9.5.0
pip==23.2.1
platformdirs==3.8.0
plumber==1.7
prompt-toolkit==3.0.38
protobuf==4.21.12
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
rabbitmq==0.2.0
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.10.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
stevedore==5.1.0
tablib==3.5.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==5.0
zope.proxy==5.0.0
